+++ R/R6Class-BipartiteSBM_sampler.R
@@ -53,6 +53,8 @@ BipartiteSBM_sampler <-
       nbBlocks    = function(value) {sapply(private$pi, length)},
       #' @field nbDyads number of dyads (potential edges in the network)
       nbDyads     = function(value) {private$dim[1] * private$dim[2]},
+      #' @field nbConnectParam number of parameter used for the connectivity
+      nbConnectParam = function(value) {self$nbBlocks^2},
       #' @field memberships list of size 2: vector of memberships in row, in column.
       memberships = function(value) {lapply(private$Z, as_clustering)},
       #' @field indMemberships list of 2 matrix for clustering memberships