+++ R/R6Class-SBM.R
@@ -109,7 +109,7 @@ SBM <- # this virtual class is the mother of all subtypes of SBM (Simple or Bipa
         cat("=====================================================================\n")
         cat("* Useful fields \n")
         cat("  $dimension, $modelName, $nbNodes, $nbBlocks, $nbCovariates, $nbDyads\n")
-        cat("  $blockProp, $connectParam, covarParam, $covarList, $covarEffect, $dimLabels \n")
+        cat("  $blockProp, $connectParam, $covarParam, $covarList, $covarEffect, $dimLabels \n")
       },
       #' @description print method
       print = function() self$show()
@@ -140,8 +140,10 @@ SBM <- # this virtual class is the mother of all subtypes of SBM (Simple or Bipa
       covarParam   = function(value) {if (missing(value)) return(private$beta) else private$beta <- value},
       #' @field covarList list of matrices of covariates
       covarList    = function(value) {if (missing(value)) return(private$X) else private$X <- value},
+      #' @field covarArray the array of covariates
+      covarArray   = function(value) {simplify2array(private$X)},
       #' @field covarEffect effect of covariates
-      covarEffect  = function(value) {if (self$nbCovariates > 0) return(roundProduct(simplify2array(private$X), private$beta)) else return(numeric(0))},
+      covarEffect  = function(value) {if (self$nbCovariates > 0) return(roundProduct(private$X, private$beta)) else return(numeric(0))},
       #' @field netMatrix the matrix (adjacency or incidence) encoding the network
       netMatrix    = function(value) {if (missing(value)) return(private$Y) else private$Y <- value},
       #' @field expectation expected values of connection under the currently adjusted model